Reduce too many levels of indirection in fread (#2389)
Previously we had `fread()` function in fread.py creating an object of class `GenericReader`, which in turn called C++ `core.gread()` function which did the actual reading.

Now the first 2 steps are removed, and `core.gread()` became the new `fread` function. The file "fread.py" moved to "utils/fread.py" after heavy editing -- it is no longer useful for any external purpose, and eventually will be removed once we replace the remaining code with C++ equivalents.

WIP for #1832
